Safari-611.1.10/LayoutTests/fast/events/touch/ios/touch-event-radius-expected.txt

PASS: radiusX is 56.8
PASS: radiusY is 56.8
